Refactor: Seperated cli logic and main.go
