+++ .gitignore
@@ -1,1 +1,1 @@
-gator
+/gator
