+++ Makefile
@@ -1,2 +1,4 @@
 build:
-	  go build -o gator .
+	  go build -o gator cmd/gator/main.go
+clean:
+		rm ./gator
